refactor: restore midi-project-list filename
